fix for pushing the player up slopes
