+++ Assets/Scripts/Physics/MovableBoxController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e7a62f332074e2e4abb64b6b077b8110
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
