--- Assets/Scripts/Player/PlayerController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 072db7d557d64d35bc2f5911428f5bc2
-timeCreated: 1684671705
